Tests SoundEffects and Displays

+++ Mekhy_Engine/Displays.py
@@ -5,6 +5,7 @@ import MachineVision
 import numpy as np
 import cv2
 import os
+import random
 display_height = 480
 display_width = 800
 display_rotation = 30
@@ -21,7 +22,7 @@ eye_scared = cv2.imread('../Eyes/eye_scared.png', cv2.COLOR_BGR2RGB)
 eye_angry = cv2.imread('../Eyes/eye_angry.png', cv2.COLOR_BGR2RGB)
 eye_disgusted = cv2.imread('../Eyes/eye_disgusted.png', cv2.COLOR_BGR2RGB)
 eye_heart = cv2.imread('../Eyes/eye_heart.png', cv2.COLOR_BGR2RGB)
-eye_hypnotic = cv2.imread('../Eyes/eye_hypnotic.png', cv2.COLOR_BGR2RGB)
+eye_hypnotic = cv2.imread('../Eyes/eye_hipnotic.png', cv2.COLOR_BGR2RGB)
 eye_sexy = cv2.imread('../Eyes/eye_sexy.png', cv2.COLOR_BGR2RGB)
 eye_demonic = cv2.imread('../Eyes/eye_demonic.png', cv2.COLOR_BGR2RGB)
 mask_neutral = cv2.VideoCapture('../Eyes/mask_neutral.mp4')
@@ -31,7 +32,7 @@ mask_scared = cv2.VideoCapture('../Eyes/mask_scared.mp4')
 mask_angry = cv2.VideoCapture('../Eyes/mask_angry.mp4')
 mask_disgusted = cv2.VideoCapture('../Eyes/mask_disgusted.mp4')
 mask_heart = cv2.VideoCapture('../Eyes/mask_heart.mp4')
-mask_hypnotic = cv2.VideoCapture('../Eyes/mask_hypnotic.mp4')
+mask_hypnotic = cv2.VideoCapture('../Eyes/mask_hipnotic.mp4')
 mask_sexy = cv2.VideoCapture('../Eyes/mask_sexy.mp4')
 mask_demonic = cv2.VideoCapture('../Eyes/mask_demonic.mp4')
 playingvideo = False
@@ -69,31 +70,48 @@ def rotateFrame(frame):
     frame = np.concatenate((firstHalf, secondHalf), axis = 1)
     return frame
 
-def composeFrame(eye, mask, displacement_eye):
-    frame = mask.copy()
+def composeFrame(eye, frame, displacement_eye):
     lefteye_center_x = int(left_eye_center[0] + (eye_radius_horizontal*displacement_eye[0]))
     lefteye_center_y = int(left_eye_center[1] + (eye_radius_vertical*displacement_eye[1]))
     righteye_center_x = int(right_eye_center[0] + (eye_radius_horizontal*displacement_eye[0]))
     righteye_center_y = int(right_eye_center[1] + (eye_radius_vertical*displacement_eye[1]))
     eyes = composeEyes(frame, eye, (lefteye_center_x, lefteye_center_y), (righteye_center_x, righteye_center_y))
-    whiteregion = np.where((frame > 125).all(axis = 2))
+    whiteregion = np.where((frame > 90).all(axis = 2))
     frame[whiteregion] = eyes[whiteregion]
     frame = rotateFrame(frame)
     return frame
 
-def PlayVideo(file_name):
+def ManageWindows():
+    screen = Wnck.Screen.get_default()
+    screen.force_update()
+    for window in screen.get_windows():
+        window_name = window.get_name()
+        if "Eyes" in window_name:
+            window.maximize()
+            window.set_geometry(Wnck.WindowGravity.STATIC, Wnck.WindowMoveResizeMask.X, 0, 0, display_width, display_height)
+            window.set_geometry(Wnck.WindowGravity.STATIC, Wnck.WindowMoveResizeMask.Y, 0, 0, display_width, display_height)
+        elif "Terminal" in window_name:
+            window.minimize()
+
+def PlayVideo(file_name, remove_file=True):
     global playingvideo
     playingvideo = True
     cap = cv2.VideoCapture(file_name)
     while(cap.isOpened() and playingvideo):
         ret, frame = cap.read()
         if ret:
-            cv2.imshow('Eyes', frame)
+            resized = cv2.resize(frame, (display_width, display_height), interpolation = cv2.INTER_AREA)
+            duplicated = cv2.hconcat([resized, resized])
+            cv2.imshow('Eyes', duplicated)
+            if cv2.waitKey(25) & 0xFF == ord('q'):
+                break
         else:
             break
+        ManageWindows()
     cap.release()
-    os.remove(file_name)
     playingvideo = False
+    if remove_file:
+        os.remove(file_name)
 
 def GraphicsRefresh(expression, displacement_eye):
     if expression == 0:
@@ -129,14 +147,18 @@ def GraphicsRefresh(expression, displacement_eye):
     if not playingvideo:
         ret, frame = mask.read()
         if ret:
-            frame = composeFrame(eye, mask, displacement_eye)
-            cv2.imshow('frame', frame)
+            frame = composeFrame(eye, frame, displacement_eye)
+            cv2.imshow('Eyes', frame)
         else:
             mask.set(cv2.CAP_PROP_POS_FRAMES, 0)
-    screen = Wnck.Screen.get_default()
-    screen.force_update()
-    for window in screen.get_windows():
-        if "Eyes" in window.get_name():
-            window.maximize()
-        elif "Terminal" in window.get_name():
-            window.minimize()
+    ManageWindows()
+
+if __name__ == '__main__':
+    PlayVideo('resources/small.mp4', False)
+    for exp in range(10):
+        x = random.uniform(-1, 1)
+        y = random.uniform(-0.3, 0.3)
+        for n in range(50):
+            GraphicsRefresh(exp, (x, y))
+            if cv2.waitKey(25) & 0xFF == ord('q'):
+                break
