+++ Mekhy_Engine/SoundEffects.py
@@ -1,4 +1,5 @@
 import os, subprocess
+import time
 import pygame
 pygame.mixer.init()
 pygame.init()
@@ -10,7 +11,7 @@ def StopSound():
     pygame.mixer.music.stop()
     pygame.mixer.stop()
 
-def PlayOnDemand(file_name):
+def PlayOnDemand(file_name, remove_file=True):
     extension = file_name.split('.')[1]
     if extension in ('mp3', 'xm'):
         pygame.mixer.music.load(file_name)
@@ -25,4 +26,15 @@ def PlayOnDemand(file_name):
         else:
             sound = pygame.mixer.Sound(file_name)
             sound.play()
-    os.remove(file_name)
+    if remove_file:
+        os.remove(file_name)
+
+if __name__ == '__main__':
+    PlayBootSound()
+    time.sleep(5)
+    PlayOnDemand('resources/sfx_WINDOWSXPBOOT.wav', False)
+    time.sleep(2)
+    StopSound()
+    PlayOnDemand('resources/file_example_MP3_700KB.mp3', False)
+    time.sleep(5)
+    StopSound()
